% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_delete_api_key}
\alias{apigateway_delete_api_key}
\title{Deletes the ApiKey resource}
\usage{
apigateway_delete_api_key(apiKey)
}
\arguments{
\item{apiKey}{[required] [Required] The identifier of the ApiKey resource to be deleted.}
}
\description{
Deletes the ApiKey resource.
}
\section{Request syntax}{
\preformatted{svc$delete_api_key(
  apiKey = "string"
)
}
}

\keyword{internal}
