% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_delete_documentation_version}
\alias{apigateway_delete_documentation_version}
\title{Delete documentation version}
\usage{
apigateway_delete_documentation_version(restApiId, documentationVersion)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{documentationVersion}{[required] [Required] The version identifier of a to-be-deleted documentation
snapshot.}
}
\description{
Delete documentation version
}
\section{Request syntax}{
\preformatted{svc$delete_documentation_version(
  restApiId = "string",
  documentationVersion = "string"
)
}
}

\keyword{internal}
