% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_base_path_mappings}
\alias{apigateway_get_base_path_mappings}
\title{Represents a collection of BasePathMapping resources}
\usage{
apigateway_get_base_path_mappings(domainName, position, limit)
}
\arguments{
\item{domainName}{[required] [Required] The domain name of a BasePathMapping resource.}

\item{position}{The current pagination position in the paged result set.}

\item{limit}{The maximum number of returned results per page. The default value is 25
and the maximum value is 500.}
}
\description{
Represents a collection of BasePathMapping resources.
}
\section{Request syntax}{
\preformatted{apigateway$get_base_path_mappings(
  domainName = "string",
  position = "string",
  limit = 123
)
}
}

\keyword{internal}
