% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_documentation_version}
\alias{apigateway_get_documentation_version}
\title{Get documentation version}
\usage{
apigateway_get_documentation_version(restApiId, documentationVersion)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{documentationVersion}{[required] [Required] The version identifier of the to-be-retrieved documentation
snapshot.}
}
\description{
Get documentation version
}
\section{Request syntax}{
\preformatted{svc$get_documentation_version(
  restApiId = "string",
  documentationVersion = "string"
)
}
}

\keyword{internal}
