% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_sdk_types}
\alias{apigateway_get_sdk_types}
\title{Get sdk types}
\usage{
apigateway_get_sdk_types(position, limit)
}
\arguments{
\item{position}{The current pagination position in the paged result set.}

\item{limit}{The maximum number of returned results per page. The default value is 25
and the maximum value is 500.}
}
\description{
Get sdk types
}
\section{Request syntax}{
\preformatted{svc$get_sdk_types(
  position = "string",
  limit = 123
)
}
}

\keyword{internal}
