% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_get_stage}
\alias{apigateway_get_stage}
\title{Gets information about a Stage resource}
\usage{
apigateway_get_stage(restApiId, stageName)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{stageName}{[required] [Required] The name of the Stage resource to get information about.}
}
\description{
Gets information about a Stage resource.
}
\section{Request syntax}{
\preformatted{svc$get_stage(
  restApiId = "string",
  stageName = "string"
)
}
}

\keyword{internal}
