% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_update_api_key}
\alias{apigateway_update_api_key}
\title{Changes information about an ApiKey resource}
\usage{
apigateway_update_api_key(apiKey, patchOperations)
}
\arguments{
\item{apiKey}{[required] [Required] The identifier of the ApiKey resource to be updated.}

\item{patchOperations}{A list of update operations to be applied to the specified resource and
in the order specified in this list.}
}
\description{
Changes information about an ApiKey resource.
}
\section{Request syntax}{
\preformatted{apigateway$update_api_key(
  apiKey = "string",
  patchOperations = list(
    list(
      op = "add"|"remove"|"replace"|"move"|"copy"|"test",
      path = "string",
      value = "string",
      from = "string"
    )
  )
)
}
}

\keyword{internal}
