% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_update_authorizer}
\alias{apigateway_update_authorizer}
\title{Updates an existing Authorizer resource}
\usage{
apigateway_update_authorizer(restApiId, authorizerId, patchOperations)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{authorizerId}{[required] [Required] The identifier of the Authorizer resource.}

\item{patchOperations}{A list of update operations to be applied to the specified resource and
in the order specified in this list.}
}
\description{
Updates an existing Authorizer resource.
}
\details{
::: {.seeAlso}
\href{https://docs.aws.amazon.com/cli/latest/reference/apigateway/update-authorizer.html}{AWS CLI}
:::
}
\section{Request syntax}{
\preformatted{svc$update_authorizer(
  restApiId = "string",
  authorizerId = "string",
  patchOperations = list(
    list(
      op = "add"|"remove"|"replace"|"move"|"copy"|"test",
      path = "string",
      value = "string",
      from = "string"
    )
  )
)
}
}

\keyword{internal}
