% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigateway_operations.R
\name{apigateway_update_usage_plan}
\alias{apigateway_update_usage_plan}
\title{Updates a usage plan of a given plan Id}
\usage{
apigateway_update_usage_plan(usagePlanId, patchOperations)
}
\arguments{
\item{usagePlanId}{[required] [Required] The Id of the to-be-updated usage plan.}

\item{patchOperations}{A list of update operations to be applied to the specified resource and
in the order specified in this list.}
}
\description{
Updates a usage plan of a given plan Id.
}
\section{Request syntax}{
\preformatted{svc$update_usage_plan(
  usagePlanId = "string",
  patchOperations = list(
    list(
      op = "add"|"remove"|"replace"|"move"|"copy"|"test",
      path = "string",
      value = "string",
      from = "string"
    )
  )
)
}
}

\keyword{internal}
