% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_get_api_mapping}
\alias{apigatewayv2_get_api_mapping}
\title{The API mapping}
\usage{
apigatewayv2_get_api_mapping(ApiMappingId, DomainName)
}
\arguments{
\item{ApiMappingId}{[required] The API mapping identifier.}

\item{DomainName}{[required] The domain name.}
}
\description{
The API mapping.
}
\section{Request syntax}{
\preformatted{svc$get_api_mapping(
  ApiMappingId = "string",
  DomainName = "string"
)
}
}

\keyword{internal}
