% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apigatewayv2_operations.R
\name{apigatewayv2_get_model_template}
\alias{apigatewayv2_get_model_template}
\title{Gets a model template}
\usage{
apigatewayv2_get_model_template(ApiId, ModelId)
}
\arguments{
\item{ApiId}{[required] The API identifier.}

\item{ModelId}{[required] The model ID.}
}
\description{
Gets a model template.
}
\section{Request syntax}{
\preformatted{svc$get_model_template(
  ApiId = "string",
  ModelId = "string"
)
}
}

\keyword{internal}
