% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appmesh_operations.R
\name{appmesh_delete_mesh}
\alias{appmesh_delete_mesh}
\title{Deletes an existing service mesh}
\usage{
appmesh_delete_mesh(meshName)
}
\arguments{
\item{meshName}{[required] The name of the service mesh to delete.}
}
\description{
Deletes an existing service mesh.
}
\details{
You must delete all resources (virtual services, routes, virtual
routers, and virtual nodes) in the service mesh before you can delete
the mesh itself.
}
\section{Request syntax}{
\preformatted{svc$delete_mesh(
  meshName = "string"
)
}
}

\keyword{internal}
