% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_distribution_config}
\alias{cloudfront_get_distribution_config}
\title{Get the configuration information about a distribution}
\usage{
cloudfront_get_distribution_config(Id)
}
\arguments{
\item{Id}{[required] The distribution's ID.}
}
\description{
Get the configuration information about a distribution.
}
\section{Request syntax}{
\preformatted{cloudfront$get_distribution_config(
  Id = "string"
)
}
}

\keyword{internal}
