% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_public_key}
\alias{cloudfront_get_public_key}
\title{Get the public key information}
\usage{
cloudfront_get_public_key(Id)
}
\arguments{
\item{Id}{[required] Request the ID for the public key.}
}
\description{
Get the public key information.
}
\section{Request syntax}{
\preformatted{cloudfront$get_public_key(
  Id = "string"
)
}
}

\keyword{internal}
