% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudfront_operations.R
\name{cloudfront_get_public_key_config}
\alias{cloudfront_get_public_key_config}
\title{Return public key configuration informaation}
\usage{
cloudfront_get_public_key_config(Id)
}
\arguments{
\item{Id}{[required] Request the ID for the public key configuration.}
}
\description{
Return public key configuration informaation
}
\section{Request syntax}{
\preformatted{svc$get_public_key_config(
  Id = "string"
)
}
}

\keyword{internal}
