% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_delete_direct_connect_gateway_association_proposal}
\alias{directconnect_delete_direct_connect_gateway_association_proposal}
\title{Deletes the association proposal request between the specified Direct
Connect gateway and virtual private gateway}
\usage{
directconnect_delete_direct_connect_gateway_association_proposal(
  proposalId)
}
\arguments{
\item{proposalId}{[required] The ID of the proposal.}
}
\description{
Deletes the association proposal request between the specified Direct
Connect gateway and virtual private gateway.
}
\section{Request syntax}{
\preformatted{svc$delete_direct_connect_gateway_association_proposal(
  proposalId = "string"
)
}
}

\keyword{internal}
