% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_describe_connections}
\alias{directconnect_describe_connections}
\title{Displays the specified connection or all connections in this Region}
\usage{
directconnect_describe_connections(connectionId)
}
\arguments{
\item{connectionId}{The ID of the connection.}
}
\description{
Displays the specified connection or all connections in this Region.
}
\section{Request syntax}{
\preformatted{svc$describe_connections(
  connectionId = "string"
)
}
}

\keyword{internal}
