% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbv2_operations.R
\name{elbv2_add_tags}
\alias{elbv2_add_tags}
\title{Adds the specified tags to the specified Elastic Load Balancing resource}
\usage{
elbv2_add_tags(ResourceArns, Tags)
}
\arguments{
\item{ResourceArns}{[required] The Amazon Resource Name (ARN) of the resource.}

\item{Tags}{[required] The tags. Each resource can have a maximum of 10 tags.}
}
\description{
Adds the specified tags to the specified Elastic Load Balancing
resource. You can tag your Application Load Balancers, Network Load
Balancers, and your target groups.
}
\details{
Each tag consists of a key and an optional value. If a resource already
has a tag with the same key, \code{AddTags} updates its value.

To list the current tags for your resources, use DescribeTags. To remove
tags from your resources, use RemoveTags.
}
\section{Request syntax}{
\preformatted{elbv2$add_tags(
  ResourceArns = list(
    "string"
  ),
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\examples{
# This example adds the specified tags to the specified load balancer.
\donttest{elbv2$add_tags(
  ResourceArns = list(
    "arn:aws:elasticloadbalancing:us-west-2:123456789012:loadbalancer/app/my-load-balancer/5..."
  ),
  Tags = list(
    list(
      Key = "project",
      Value = "lima"
    ),
    list(
      Key = "department",
      Value = "digital-media"
    )
  )
)}

}
\keyword{internal}
