% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_delete_accelerator}
\alias{globalaccelerator_delete_accelerator}
\title{Delete an accelerator}
\usage{
globalaccelerator_delete_accelerator(AcceleratorArn)
}
\arguments{
\item{AcceleratorArn}{[required] The Amazon Resource Name (ARN) of an accelerator.}
}
\description{
Delete an accelerator. Note: before you can delete an accelerator, you
must disable it and remove all dependent resources (listeners and
endpoint groups).
}
\section{Request syntax}{
\preformatted{globalaccelerator$delete_accelerator(
  AcceleratorArn = "string"
)
}
}

\keyword{internal}
