% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_describe_listener}
\alias{globalaccelerator_describe_listener}
\title{Describe a listener}
\usage{
globalaccelerator_describe_listener(ListenerArn)
}
\arguments{
\item{ListenerArn}{[required] The Amazon Resource Name (ARN) of the listener to describe.}
}
\description{
Describe a listener.
}
\section{Request syntax}{
\preformatted{svc$describe_listener(
  ListenerArn = "string"
)
}
}

\keyword{internal}
