% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaccelerator_operations.R
\name{globalaccelerator_update_accelerator_attributes}
\alias{globalaccelerator_update_accelerator_attributes}
\title{Update the attributes for an accelerator}
\usage{
globalaccelerator_update_accelerator_attributes(AcceleratorArn,
  FlowLogsEnabled, FlowLogsS3Bucket, FlowLogsS3Prefix)
}
\arguments{
\item{AcceleratorArn}{The Amazon Resource Name (ARN) of the accelerator that you want to
update. Attribute is required.}

\item{FlowLogsEnabled}{Update whether flow logs are enabled. The default value is false. If the
value is true, \code{FlowLogsS3Bucket} and \code{FlowLogsS3Prefix} must be
specified.

For more information, see \href{https://docs.aws.amazon.com/global-accelerator/latest/dg/monitoring-global-accelerator.flow-logs.html}{Flow Logs}
in the \emph{AWS Global Accelerator Developer Guide}.}

\item{FlowLogsS3Bucket}{The name of the Amazon S3 bucket for the flow logs. Attribute is
required if \code{FlowLogsEnabled} is \code{true}. The bucket must exist and have
a bucket policy that grants AWS Global Accelerator permission to write
to the bucket.}

\item{FlowLogsS3Prefix}{Update the prefix for the location in the Amazon S3 bucket for the flow
logs. Attribute is required if \code{FlowLogsEnabled} is \code{true}. If you don't
specify a prefix, the flow logs are stored in the root of the bucket.}
}
\description{
Update the attributes for an accelerator. To see an AWS CLI example of
updating an accelerator to enable flow logs, scroll down to \strong{Example}.
}
\section{Request syntax}{
\preformatted{svc$update_accelerator_attributes(
  AcceleratorArn = "string",
  FlowLogsEnabled = TRUE|FALSE,
  FlowLogsS3Bucket = "string",
  FlowLogsS3Prefix = "string"
)
}
}

\keyword{internal}
