% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53_operations.R
\name{route53_get_health_check_count}
\alias{route53_get_health_check_count}
\title{Retrieves the number of health checks that are associated with the
current AWS account}
\usage{
route53_get_health_check_count()
}
\description{
Retrieves the number of health checks that are associated with the
current AWS account.
}
\section{Request syntax}{
\preformatted{svc$get_health_check_count()
}
}

\keyword{internal}
