% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53_operations.R
\name{route53_get_health_check_status}
\alias{route53_get_health_check_status}
\title{Gets status of a specified health check}
\usage{
route53_get_health_check_status(HealthCheckId)
}
\arguments{
\item{HealthCheckId}{[required] The ID for the health check that you want the current status for. When
you created the health check, \code{CreateHealthCheck} returned the ID in the
response, in the \code{HealthCheckId} element.

If you want to check the status of a calculated health check, you must
use the Amazon Route 53 console or the CloudWatch console. You can't
use \code{GetHealthCheckStatus} to get the status of a calculated health
check.}
}
\description{
Gets status of a specified health check.
}
\section{Request syntax}{
\preformatted{svc$get_health_check_status(
  HealthCheckId = "string"
)
}
}

\keyword{internal}
