% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53_operations.R
\name{route53_get_traffic_policy}
\alias{route53_get_traffic_policy}
\title{Gets information about a specific traffic policy version}
\usage{
route53_get_traffic_policy(Id, Version)
}
\arguments{
\item{Id}{[required] The ID of the traffic policy that you want to get information about.}

\item{Version}{[required] The version number of the traffic policy that you want to get
information about.}
}
\description{
Gets information about a specific traffic policy version.
}
\section{Request syntax}{
\preformatted{svc$get_traffic_policy(
  Id = "string",
  Version = 123
)
}
}

\keyword{internal}
