% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53_operations.R
\name{route53_get_traffic_policy_instance_count}
\alias{route53_get_traffic_policy_instance_count}
\title{Gets the number of traffic policy instances that are associated with the
current AWS account}
\usage{
route53_get_traffic_policy_instance_count()
}
\description{
Gets the number of traffic policy instances that are associated with the
current AWS account.
}
\section{Request syntax}{
\preformatted{route53$get_traffic_policy_instance_count()
}
}

\keyword{internal}
