% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53resolver_operations.R
\name{route53resolver_get_resolver_rule_policy}
\alias{route53resolver_get_resolver_rule_policy}
\title{Gets information about a resolver rule policy}
\usage{
route53resolver_get_resolver_rule_policy(Arn)
}
\arguments{
\item{Arn}{[required] The ID of the resolver rule policy that you want to get information
about.}
}
\description{
Gets information about a resolver rule policy. A resolver rule policy
specifies the Resolver operations and resources that you want to allow
another AWS account to be able to use.
}
\section{Request syntax}{
\preformatted{svc$get_resolver_rule_policy(
  Arn = "string"
)
}
}

\keyword{internal}
