% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicediscovery_operations.R
\name{servicediscovery_list_instances}
\alias{servicediscovery_list_instances}
\title{Lists summary information about the instances that you registered by
using a specified service}
\usage{
servicediscovery_list_instances(ServiceId, NextToken, MaxResults)
}
\arguments{
\item{ServiceId}{[required] The ID of the service that you want to list instances for.}

\item{NextToken}{For the first \code{ListInstances} request, omit this value.

If more than \code{MaxResults} instances match the specified criteria, you
can submit another \code{ListInstances} request to get the next group of
results. Specify the value of \code{NextToken} from the previous response in
the next request.}

\item{MaxResults}{The maximum number of instances that you want AWS Cloud Map to return in
the response to a \code{ListInstances} request. If you don't specify a value
for \code{MaxResults}, AWS Cloud Map returns up to 100 instances.}
}
\description{
Lists summary information about the instances that you registered by
using a specified service.
}
\section{Request syntax}{
\preformatted{servicediscovery$list_instances(
  ServiceId = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
